+++ redvox/common/io.py
@@ -9,8 +9,6 @@ import os.path
 import multiprocessing
 import multiprocessing.pool
 from pathlib import Path, PurePath
-import pickle
-import json
 from shutil import copy2, move
 from typing import (
     Any,